working in edit-profile v0.7 start create swager and minor changes
