+++ apps/auth/swagger/posts.swagger.ts
@@ -55,35 +55,9 @@ export function CreatePostEndpoint() {
     ApiBearerAuth(),
     ApiConsumes('multipart/form-data'),
     //***********************
-    ApiBody({
-      description: 'Create new post with photos',
-      type: 'multipart/form-data', // Указываем, что это multipart
-      required: true,
-      schema: {
-        type: 'object',
-        properties: {
-          content: {
-            type: 'string',
-            maxLength: 500,
-            description:
-              'Some content which describe photos, It`s not mandatory!',
-            example: 'Some information about the photo, this is not mandatory.',
-          },
-          files: {
-            type: 'array', // Указываем, что это массив файлов
-            items: {
-              type: 'string',
-              format: 'binary', // Указываем, что это файл
-            },
-            minItems: 1,
-            maxItems: 10,
-          },
-        },
-      },
-    }),
-    //***********************
     // ApiBody({
-    //   description: 'Create Post',
+    //   description: 'Create new post with photos',
+    //   type: 'multipart/form-data', // Указываем, что это multipart
     //   required: true,
     //   schema: {
     //     type: 'object',
@@ -91,23 +65,52 @@ export function CreatePostEndpoint() {
     //       content: {
     //         type: 'string',
     //         maxLength: 500,
-    //         description: 'Some content which describe photos, NO Mandatory!',
+    //         description:
+    //           'Some content which describe photos, It`s not mandatory!',
     //         example: 'Some information about the photo, this is not mandatory.',
     //       },
-    //       photos: {
-    //         type: 'array',
+    //       files: {
+    //         type: 'array', // Указываем, что это массив файлов
     //         items: {
     //           type: 'string',
-    //           format: 'binary',
-    //           description: 'Download photo min 1 max 10',
+    //           format: 'binary', // Указываем, что это файл
     //         },
     //         minItems: 1,
     //         maxItems: 10,
     //       },
     //     },
-    //     required: ['photos'],
     //   },
     // }),
+    //***********************
+    ApiBody({
+      description: 'Create Post',
+      required: true,
+      schema: {
+        type: 'object',
+        properties: {
+          content: {
+            type: 'string',
+            maxLength: 500,
+            description: 'Some content which describe photos, NO Mandatory!',
+            example: 'Some information about the photo, this is not mandatory.',
+          },
+          photos: {
+            type: 'array',
+            format: 'binary',
+            description:
+              'Download photo min 1 max 10, max size 2mb per photo, format .jpg, .png',
+            // items: {
+            //   type: 'string',
+            //   format: 'binary',
+            //   description: 'Download photo min 1 max 10',
+            // },
+            // minItems: 1,
+            // maxItems: 10,
+          },
+        },
+        required: ['photos'],
+      },
+    }),
     ApiResponse({
       status: 201,
       description: 'Return postId',
